refactor: group components into directories
