+++ client/js/bundles/user/listOrgas.js
@@ -13,7 +13,7 @@
 
 import { DpEditor, dpValidate } from '@demos-europe/demosplan-ui'
 import DpCreateItem from '@DpJs/components/user/DpCreateItem'
-import DpOrganisationList from '@DpJs/components/user/DpOrganisationList'
+import DpOrganisationList from '@DpJs/components/user/DpOrganisationList/DpOrganisationList'
 import { initialize } from '@DpJs/InitVue'
 import UrlPreview from '@DpJs/lib/shared/UrlPreview'
