+++ client/js/bundles/user/listUser.js
@@ -12,7 +12,7 @@
  */
 
 import DpCreateItem from '@DpJs/components/user/DpCreateItem'
-import DpUserList from '@DpJs/components/user/DpUserList'
+import DpUserList from '@DpJs/components/user/DpUserList/DpUserList'
 import DpUserListExtended from '@DpJs/components/user/DpUserListExtended'
 import { initialize } from '@DpJs/InitVue'
 import UserFormFields from '@DpJs/store/user/UserFormFields'
